+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Replace defaultProps with JavaScript default parameters
+
 ## [v1.1.4] - 2020-08-03
 
 ### Changed